Refactor Time!
There was too much business logic mixed in with the CoreLocation code and I had all kinds of weird bugs and improperly handled edge cases. So I made the AppManager the CLLocationManagerDelegate and set it up to pass my custom Location objects to the Commute Manager.

The CommuteManager now has a list of endpoints which are wrappers around CLCircularRegion with the addition of a pair of CommuteSchedules (one for entry and one for exit). It also manages an Active Commute in the store rather than just in memory.